refactor: remove redundant numeric casts (Sonar S1905)
Strip casts whose target equals the operand's existing type after
auto-widening — value-preserving no-ops flagged by SonarCloud S1905.
Overflow-guard sites kept correct: the surviving operand is already
long, so multiplication/subtraction stays long-arithmetic.

Generated fbs/ sources left untouched (regenerated).

+++ cli/src/main/java/io/github/dfa1/vortex/cli/tui/VortexInspectorTui.java
@@ -839,7 +839,7 @@ public final class VortexInspectorTui {
             Layout layout = node.layout();
             int segIdx = layout.segments().getFirst();
             SegmentSpec spec = tree.segmentSpecs().get(segIdx);
-            int wanted = (int) Math.min((long) HEX_PREVIEW_BYTES, spec.length());
+            int wanted = (int) Math.min(HEX_PREVIEW_BYTES, spec.length());
             if (wanted <= 0) {
                 return new byte[0];
             }
